fix: 单舱测试改为真实执行 + 移除表单红色必填标记
1. 单舱测试（test-run）从纯配置校验改为真实执行舱逻辑：
   - 提取 CapsuleExecutor 静态类，封装所有舱执行逻辑
   - WorkflowRunWorker 和 Controller 共享同一份执行代码
   - 测试按钮直接调用对应舱类型的执行引擎，返回真实产物和日志
   - 移除必填字段校验阻断，让执行引擎自然报错

2. 前端测试结果展示重构：
   - 显示执行日志、产物内容（可展开）、错误信息
   - 测试按钮增加 loading 状态
   - 移除旧的 configValidation 逐字段校验展示

3. 移除所有红色必填标记（* 和 "必需" 文字）：
   - CapsuleConfigForm 表单字段不再显示红色星号
   - VariablesSection 变量字段同样移除
   - 输入插槽的 "必需" 改为 muted 颜色
